+++ components/clarinet-cli/src/lsp/native_bridge.rs
@@ -1,8 +1,8 @@
 use super::utils;
 use crate::lsp::clarity_diagnostics_to_tower_lsp_type;
 use clarity_lsp::backend::{
-    process_notification, process_request, EditorStateInput, LspNotification,
-    LspNotificationResponse, LspRequest, LspRequestResponse,
+    process_mutating_request, process_notification, process_request, EditorStateInput,
+    LspNotification, LspNotificationResponse, LspRequest, LspRequestResponse,
 };
 use clarity_lsp::lsp_types::{
     DocumentSymbolParams, DocumentSymbolResponse, GotoDefinitionParams, GotoDefinitionResponse,
@@ -54,8 +54,15 @@ pub async fn start_language_server(
             },
             i if i == requests_oper => match oper.recv(&request_rx) {
                 Ok(request) => {
-                    let request_response = process_request(request, &editor_state);
-                    let _ = response_tx.send(LspResponse::Request(request_response));
+                    let request_result = match request {
+                        LspRequest::Initialize(_) => {
+                            process_mutating_request(request, &mut editor_state)
+                        }
+                        _ => process_request(request, &editor_state),
+                    };
+                    if let Ok(response) = request_result {
+                        let _ = response_tx.send(LspResponse::Request(response));
+                    }
                 }
                 Err(_e) => {
                     continue;