+++ components/clarity-lsp/src/common/backend.rs
@@ -1,14 +1,13 @@
 use crate::lsp_types::MessageType;
 use crate::state::{build_state, EditorState, ProtocolState};
-use crate::types::{CompletionItemKind, InsertTextFormat};
 use crate::utils::get_contract_location;
 use clarinet_files::{FileAccessor, FileLocation, ProjectManifest};
 use clarity_repl::clarity::diagnostic::Diagnostic;
 use clarity_repl::repl::ContractDeployer;
 use lsp_types::{
-    CompletionItem, CompletionParams, DocumentSymbol, DocumentSymbolParams, Documentation,
-    GotoDefinitionParams, Hover, HoverParams, InitializeParams, InitializeResult, Location,
-    MarkupContent, MarkupKind, SignatureHelp, SignatureHelpParams,
+    CompletionItem, CompletionParams, DocumentSymbol, DocumentSymbolParams, GotoDefinitionParams,
+    Hover, HoverParams, InitializeParams, InitializeResult, Location, SignatureHelp,
+    SignatureHelpParams,
 };
 use serde::{Deserialize, Serialize};
 use std::sync::{Arc, RwLock};
@@ -167,7 +166,10 @@ pub async fn process_notification(
                         }?
                         .contracts_settings
                         .get(&contract_location)
-                        .ok_or("contract not found in manifest")?
+                        .ok_or(format!(
+                            "No Clarinet.toml is associated to the contract {}",
+                            &contract_location.get_file_name().unwrap_or_default()
+                        ))?
                         .clone()
                         .clarity_version
                     }
@@ -261,147 +263,66 @@ pub async fn process_notification(
 
 #[derive(Debug, Serialize, Deserialize)]
 pub enum LspRequest {
-    Initialize(InitializeParams),
     Completion(CompletionParams),
     SignatureHelp(SignatureHelpParams),
     Definition(GotoDefinitionParams),
     Hover(HoverParams),
     DocumentSymbol(DocumentSymbolParams),
+    Initialize(InitializeParams),
 }
 
 #[derive(Debug, PartialEq, Deserialize, Serialize)]
 pub enum LspRequestResponse {
-    Initialize(InitializeResult),
     CompletionItems(Vec<CompletionItem>),
     SignatureHelp(Option<SignatureHelp>),
     Definition(Option<Location>),
     DocumentSymbol(Vec<DocumentSymbol>),
     Hover(Option<Hover>),
+    Initialize(InitializeResult),
 }
 
-pub fn process_request(command: LspRequest, editor_state: &EditorStateInput) -> LspRequestResponse {
+pub fn process_request(
+    command: LspRequest,
+    editor_state: &EditorStateInput,
+) -> Result<LspRequestResponse, String> {
     match command {
-        LspRequest::Initialize(params) => {
-            let initialization_options: InitializationOptions = params
-                .initialization_options
-                .and_then(|o| serde_json::from_str(o.as_str()?).ok())
-                .expect("failed to parse initialization options");
-
-            LspRequestResponse::Initialize(InitializeResult {
-                server_info: None,
-                capabilities: get_capabilities(&initialization_options),
-            })
-        }
-
         LspRequest::Completion(params) => {
             let file_url = params.text_document_position.text_document.uri;
+            let position = params.text_document_position.position;
+
             let contract_location = match get_contract_location(&file_url) {
                 Some(contract_location) => contract_location,
-                None => return LspRequestResponse::CompletionItems(vec![]),
+                None => return Ok(LspRequestResponse::CompletionItems(vec![])),
             };
-            let mut completion_items_src = match editor_state
-                .try_read(|es| es.get_completion_items_for_contract(&contract_location))
+
+            let completion_items = match editor_state
+                .try_read(|es| es.get_completion_items_for_contract(&contract_location, &position))
             {
                 Ok(result) => result,
-                Err(_) => return LspRequestResponse::CompletionItems(vec![]),
+                Err(_) => return Ok(LspRequestResponse::CompletionItems(vec![])),
             };
 
-            let mut completion_items = vec![];
-            // Little big detail: should we wrap the inserted_text with braces?
-            let should_wrap = {
-                // let line = params.text_document_position.position.line;
-                // let char = params.text_document_position.position.character;
-                // let doc = params.text_document_position.text_document.uri;
-                //
-                // TODO(lgalabru): from there, we'd need to get the prior char
-                // and see if a parenthesis was opened. If not, we need to wrap.
-                // The LSP would need to update its local document cache, via
-                // the did_change method.
-                true
-            };
-
-            if should_wrap {
-                for mut item in completion_items_src.drain(..) {
-                    match item.kind {
-                        CompletionItemKind::Event
-                        | CompletionItemKind::Function
-                        | CompletionItemKind::Module
-                        | CompletionItemKind::Class => {
-                            item.insert_text =
-                                Some(format!("({})", item.insert_text.take().unwrap()));
-                        }
-                        _ => {}
-                    }
-
-                    let kind = match item.kind {
-                        CompletionItemKind::Class => lsp_types::CompletionItemKind::CLASS,
-                        CompletionItemKind::Event => lsp_types::CompletionItemKind::EVENT,
-                        CompletionItemKind::Field => lsp_types::CompletionItemKind::FIELD,
-                        CompletionItemKind::Function => lsp_types::CompletionItemKind::FUNCTION,
-                        CompletionItemKind::Module => lsp_types::CompletionItemKind::MODULE,
-                        CompletionItemKind::TypeParameter => {
-                            lsp_types::CompletionItemKind::TYPE_PARAMETER
-                        }
-                    };
-
-                    let insert_text_format = match item.insert_text_format {
-                        InsertTextFormat::PlainText => lsp_types::InsertTextFormat::PLAIN_TEXT,
-                        InsertTextFormat::Snippet => lsp_types::InsertTextFormat::SNIPPET,
-                    };
-
-                    let completion_item = CompletionItem {
-                        label: item.label.clone(),
-                        kind: Some(kind),
-                        detail: item.detail.take(),
-                        documentation: item.markdown_documentation.take().and_then(|doc| {
-                            Some(Documentation::MarkupContent(MarkupContent {
-                                kind: MarkupKind::Markdown,
-                                value: doc,
-                            }))
-                        }),
-                        deprecated: None,
-                        preselect: None,
-                        sort_text: None,
-                        filter_text: None,
-                        insert_text: item.insert_text.take(),
-                        insert_text_format: Some(insert_text_format),
-                        insert_text_mode: None,
-                        text_edit: None,
-                        additional_text_edits: None,
-                        command: Some(lsp_types::Command {
-                            title: "triggerParameterHints".into(),
-                            command: "editor.action.triggerParameterHints".into(),
-                            arguments: None,
-                        }),
-                        commit_characters: None,
-                        data: None,
-                        tags: None,
-                    };
-                    completion_items.push(completion_item);
-                }
-            }
-
-            LspRequestResponse::CompletionItems(completion_items)
+            Ok(LspRequestResponse::CompletionItems(completion_items))
         }
 
         LspRequest::Definition(params) => {
             let file_url = params.text_document_position_params.text_document.uri;
             let contract_location = match get_contract_location(&file_url) {
                 Some(contract_location) => contract_location,
-                None => return LspRequestResponse::Definition(None),
+                None => return Ok(LspRequestResponse::Definition(None)),
             };
             let position = params.text_document_position_params.position;
             let location = editor_state
                 .try_read(|es| es.get_definition_location(&contract_location, &position))
                 .unwrap_or_default();
-            LspRequestResponse::Definition(location)
+            Ok(LspRequestResponse::Definition(location))
         }
 
         LspRequest::SignatureHelp(params) => {
             let file_url = params.text_document_position_params.text_document.uri;
             let contract_location = match get_contract_location(&file_url) {
                 Some(contract_location) => contract_location,
-                None => return LspRequestResponse::Definition(None),
+                None => return Ok(LspRequestResponse::SignatureHelp(None)),
             };
             let position = params.text_document_position_params.position;
 
@@ -417,32 +338,64 @@ pub fn process_request(command: LspRequest, editor_state: &EditorStateInput) -> 
                     es.get_signature_help(&contract_location, &position, active_signature)
                 })
                 .unwrap_or_default();
-            LspRequestResponse::SignatureHelp(signature)
+            Ok(LspRequestResponse::SignatureHelp(signature))
         }
 
         LspRequest::DocumentSymbol(params) => {
             let file_url = params.text_document.uri;
             let contract_location = match get_contract_location(&file_url) {
                 Some(contract_location) => contract_location,
-                None => return LspRequestResponse::DocumentSymbol(vec![]),
+                None => return Ok(LspRequestResponse::DocumentSymbol(vec![])),
             };
             let document_symbols = editor_state
                 .try_read(|es| es.get_document_symbols_for_contract(&contract_location))
                 .unwrap_or_default();
-            LspRequestResponse::DocumentSymbol(document_symbols)
+            Ok(LspRequestResponse::DocumentSymbol(document_symbols))
         }
 
         LspRequest::Hover(params) => {
             let file_url = params.text_document_position_params.text_document.uri;
             let contract_location = match get_contract_location(&file_url) {
                 Some(contract_location) => contract_location,
-                None => return LspRequestResponse::Hover(None),
+                None => return Ok(LspRequestResponse::Hover(None)),
             };
             let position = params.text_document_position_params.position;
             let hover_data = editor_state
                 .try_read(|es| es.get_hover_data(&contract_location, &position))
                 .unwrap_or_default();
-            LspRequestResponse::Hover(hover_data)
+            Ok(LspRequestResponse::Hover(hover_data))
         }
+        _ => Err(format!("Unexpected command: {:?}", &command)),
+    }
+}
+
+// lsp requests are not supposed to mut the editor_state (only the notifications do)
+// this is to ensure there is no concurrency between notifications and requests to
+// acquire write lock on the editor state in a wasm context
+// except for the Initialize request, which is the first interaction between the client and the server
+// and can therefore safely acquire write lock on the editor state
+pub fn process_mutating_request(
+    command: LspRequest,
+    editor_state: &mut EditorStateInput,
+) -> Result<LspRequestResponse, String> {
+    match command {
+        LspRequest::Initialize(params) => {
+            let initialization_options: InitializationOptions = params
+                .initialization_options
+                .and_then(|o| serde_json::from_str(o.as_str()?).ok())
+                .expect("failed to parse initialization options");
+
+            match editor_state.try_write(|es| es.settings = initialization_options.clone()) {
+                Ok(_) => Ok(LspRequestResponse::Initialize(InitializeResult {
+                    server_info: None,
+                    capabilities: get_capabilities(&initialization_options),
+                })),
+                Err(err) => Err(err),
+            }
+        }
+        _ => Err(format!(
+            "Unexpected command: {:?}, should not not mutate state",
+            &command
+        )),
     }
 }